+++ src/components/account/accountReview/accountReviewModal/addPhotoFile.js
@@ -1,15 +1,35 @@
-import React from "react";
+import React, { useState, useRef, useCallback } from "react";
 import classes from "./addPhotoFile.module.css";
 
 function AddPhotoFile() {
+  //파일 미리볼 url을 저장해줄 state
+  const [fileImage, setFileImage] = useState("");
+  // 파일 저장
+  const saveFileImage = e => {
+    setFileImage(URL.createObjectURL(e.target.files[0]));
+  };
+  // 파일 삭제
+  const deleteFileImage = () => {
+    URL.revokeObjectURL(fileImage);
+    setFileImage("");
+  };
+
   return (
     <div className={classes.wrapper}>
       <p className={classes.title}>사진첨부(선택)</p>
       <p className={classes.description}>사진을 첨부해주세요. (최대 1장)</p>
-      <div className={classes.filebox}>
-        <label for="addPhoto">사진첨부하기</label>
-        <input id="addPhoto" type="file" />
+      <div className={classes.imgDiv}>
+        {fileImage && (
+          <img alt="sample" src={fileImage} style={{ margin: "auto" }} />
+        )}
       </div>
+      <input
+        className={classes.filebox}
+        name="imgUpload"
+        type="file"
+        accept="image/*"
+        onChange={saveFileImage}
+      />
     </div>
   );
 }